Validação dos dados de perfil de usuário

+++ app/Models/User.php
@@ -55,14 +55,19 @@ class User extends Authenticatable implements JWTSubject
     public function updateWithProfile(array $data): User
     {
         try{
-            UserProfile::uploadPhoto($data['photo']);
+            if(isset($data['photo'])){
+                UserProfile::uploadPhoto($data['photo']);
+            }
+
             \DB::beginTransaction();
             $this->fill($data);
             $this->save();
             UserProfile::saveProfile($this, $data);
             \DB::commit();
         }catch (\Exception $e){
-            UserProfile::deleteFile($data['photo']);
+            if(isset($data['photo'])) {
+                UserProfile::deleteFile($data['photo']);
+            }
             \DB::rollBack();
             throw $e;
         }
